bindinfor: refacter the matcher
