+++ pkg/bindinfo/BUILD.bazel
@@ -14,6 +14,7 @@ go_library(
     importpath = "github.com/pingcap/tidb/pkg/bindinfo",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/bindinfo/norm",
         "//pkg/kv",
         "//pkg/metrics",
         "//pkg/parser",
@@ -63,6 +64,7 @@ go_test(
     shard_count = 50,
     deps = [
         "//pkg/bindinfo/internal",
+        "//pkg/bindinfo/norm",
         "//pkg/config",
         "//pkg/domain",
         "//pkg/metrics",
