+++ pkg/bindinfo/binding_match.go
@@ -15,15 +15,13 @@
 package bindinfo
 
 import (
-	"strings"
 	"sync"
 
+	"github.com/pingcap/tidb/pkg/bindinfo/norm"
 	"github.com/pingcap/tidb/pkg/metrics"
-	"github.com/pingcap/tidb/pkg/parser"
 	"github.com/pingcap/tidb/pkg/parser/ast"
 	"github.com/pingcap/tidb/pkg/sessionctx"
 	"github.com/pingcap/tidb/pkg/util/hint"
-	utilparser "github.com/pingcap/tidb/pkg/util/parser"
 )
 
 var (
@@ -74,7 +72,7 @@ func getBindRecord(sctx sessionctx.Context, stmtNode ast.StmtNode, info *Binding
 	var fuzzyDigest string
 	var tableNames []*ast.TableName
 	if info == nil || info.TableNames == nil || info.FuzzyDigest == "" {
-		_, fuzzyDigest = NormalizeStmtForFuzzyBinding(stmtNode)
+		_, fuzzyDigest = norm.NormalizeStmtForFuzzyBinding(stmtNode)
 		tableNames = CollectTableNames(stmtNode)
 		if info != nil {
 			info.FuzzyDigest = fuzzyDigest
@@ -97,90 +95,6 @@ func getBindRecord(sctx sessionctx.Context, stmtNode ast.StmtNode, info *Binding
 		return bindRecord, metrics.ScopeGlobal
 	}
 	return nil, ""
-}
-
-func eraseLastSemicolon(stmt ast.StmtNode) {
-	sql := stmt.Text()
-	if len(sql) > 0 && sql[len(sql)-1] == ';' {
-		stmt.SetText(nil, sql[:len(sql)-1])
-	}
-}
-
-// NormalizeStmtForBinding normalizes a statement for binding.
-// Schema names will be completed automatically: `select * from t` --> `select * from db . t`.
-func NormalizeStmtForBinding(stmtNode ast.StmtNode, specifiedDB string) (normalizedStmt, exactSQLDigest string) {
-	return normalizeStmt(stmtNode, specifiedDB, false)
-}
-
-// NormalizeStmtForFuzzyBinding normalizes a statement for fuzzy matching.
-// Schema names will be eliminated automatically: `select * from db . t` --> `select * from t`.
-func NormalizeStmtForFuzzyBinding(stmtNode ast.StmtNode) (normalizedStmt, fuzzySQLDigest string) {
-	return normalizeStmt(stmtNode, "", true)
-}
-
-// NormalizeStmtForBinding normalizes a statement for binding.
-// This function skips Explain automatically, and literals in in-lists will be normalized as '...'.
-// For normal bindings, DB name will be completed automatically:
-//
-//	e.g. `select * from t where a in (1, 2, 3)` --> `select * from test.t where a in (...)`
-func normalizeStmt(stmtNode ast.StmtNode, specifiedDB string, fuzzy bool) (normalizedStmt, sqlDigest string) {
-	normalize := func(n ast.StmtNode) (normalizedStmt, sqlDigest string) {
-		eraseLastSemicolon(n)
-		var digest *parser.Digest
-		var normalizedSQL string
-		if !fuzzy {
-			normalizedSQL = utilparser.RestoreWithDefaultDB(n, specifiedDB, n.Text())
-		} else {
-			normalizedSQL = utilparser.RestoreWithoutDB(n)
-		}
-		normalizedStmt, digest = parser.NormalizeDigestForBinding(normalizedSQL)
-		return normalizedStmt, digest.String()
-	}
-
-	switch x := stmtNode.(type) {
-	case *ast.ExplainStmt:
-		// This function is only used to find bind record.
-		// For some SQLs, such as `explain select * from t`, they will be entered here many times,
-		// but some of them do not want to obtain bind record.
-		// The difference between them is whether len(x.Text()) is empty. They cannot be distinguished by stmt.restore.
-		// For these cases, we need return "" as normalize SQL and hash.
-		if len(x.Text()) == 0 {
-			return "", ""
-		}
-		switch x.Stmt.(type) {
-		case *ast.SelectStmt, *ast.DeleteStmt, *ast.UpdateStmt, *ast.InsertStmt:
-			normalizeSQL, digest := normalize(x.Stmt)
-			return normalizeSQL, digest
-		case *ast.SetOprStmt:
-			normalizeExplainSQL, _ := normalize(x)
-
-			idx := strings.Index(normalizeExplainSQL, "select")
-			parenthesesIdx := strings.Index(normalizeExplainSQL, "(")
-			if parenthesesIdx != -1 && parenthesesIdx < idx {
-				idx = parenthesesIdx
-			}
-			// If the SQL is `EXPLAIN ((VALUES ROW ()) ORDER BY 1);`, the idx will be -1.
-			if idx == -1 {
-				hash := parser.DigestNormalized(normalizeExplainSQL)
-				return normalizeExplainSQL, hash.String()
-			}
-			normalizeSQL := normalizeExplainSQL[idx:]
-			hash := parser.DigestNormalized(normalizeSQL)
-			return normalizeSQL, hash.String()
-		}
-	case *ast.SelectStmt, *ast.SetOprStmt, *ast.DeleteStmt, *ast.UpdateStmt, *ast.InsertStmt:
-		// This function is only used to find bind record.
-		// For some SQLs, such as `explain select * from t`, they will be entered here many times,
-		// but some of them do not want to obtain bind record.
-		// The difference between them is whether len(x.Text()) is empty. They cannot be distinguished by stmt.restore.
-		// For these cases, we need return "" as normalize SQL and hash.
-		if len(x.Text()) == 0 {
-			return "", ""
-		}
-		normalizedSQL, digest := normalize(x)
-		return normalizedSQL, digest
-	}
-	return "", ""
 }
 
 func fuzzyMatchBindingTableName(currentDB string, stmtTableNames, bindingTableNames []*ast.TableName) (numWildcards int, matched bool) {
