+++ pkg/bindinfo/tests/bind_test.go
@@ -22,6 +22,7 @@ import (
 
 	"github.com/pingcap/tidb/pkg/bindinfo"
 	"github.com/pingcap/tidb/pkg/bindinfo/internal"
+	"github.com/pingcap/tidb/pkg/bindinfo/norm"
 	"github.com/pingcap/tidb/pkg/domain"
 	"github.com/pingcap/tidb/pkg/parser"
 	"github.com/pingcap/tidb/pkg/parser/model"
@@ -320,7 +321,7 @@ func TestBindingSymbolList(t *testing.T) {
 	stmt, err := parser.New().ParseOneStmt("select a, b from test . t where a = 1 limit 0, 1", "", "")
 	require.NoError(t, err)
 
-	_, fuzzyDigest := bindinfo.NormalizeStmtForFuzzyBinding(stmt)
+	_, fuzzyDigest := norm.NormalizeStmtForFuzzyBinding(stmt)
 	bindData, err := dom.BindHandle().MatchGlobalBinding(tk.Session(), fuzzyDigest, bindinfo.CollectTableNames(stmt))
 	require.NoError(t, err)
 	require.NotNil(t, bindData)
@@ -367,7 +368,7 @@ func TestBindingInListWithSingleLiteral(t *testing.T) {
 	stmt, err := parser.New().ParseOneStmt("select a, b from test . t where a in (1)", "", "")
 	require.NoError(t, err)
 
-	_, fuzzyDigest := bindinfo.NormalizeStmtForFuzzyBinding(stmt)
+	_, fuzzyDigest := norm.NormalizeStmtForFuzzyBinding(stmt)
 	bindData, err := dom.BindHandle().MatchGlobalBinding(tk.Session(), fuzzyDigest, bindinfo.CollectTableNames(stmt))
 	require.NoError(t, err)
 	require.NotNil(t, bindData)
@@ -405,7 +406,7 @@ func TestBestPlanInBaselines(t *testing.T) {
 
 	stmt, _, _ := internal.UtilNormalizeWithDefaultDB(t, "select a, b from t where a = 1 limit 0, 1")
 
-	_, fuzzyDigest := bindinfo.NormalizeStmtForFuzzyBinding(stmt)
+	_, fuzzyDigest := norm.NormalizeStmtForFuzzyBinding(stmt)
 	bindData, err := dom.BindHandle().MatchGlobalBinding(tk.Session(), fuzzyDigest, bindinfo.CollectTableNames(stmt))
 	require.NoError(t, err)
 	require.NotNil(t, bindData)
@@ -441,7 +442,7 @@ func TestErrorBind(t *testing.T) {
 
 	stmt, err := parser.New().ParseOneStmt("select * from test . t where i > ?", "", "")
 	require.NoError(t, err)
-	_, fuzzyDigest := bindinfo.NormalizeStmtForFuzzyBinding(stmt)
+	_, fuzzyDigest := norm.NormalizeStmtForFuzzyBinding(stmt)
 	bindData, err := dom.BindHandle().MatchGlobalBinding(tk.Session(), fuzzyDigest, bindinfo.CollectTableNames(stmt))
 	require.NoError(t, err)
 	require.NotNil(t, bindData)
@@ -500,7 +501,7 @@ func TestHintsSetID(t *testing.T) {
 	// Verify the added Binding contains ID with restored query block.
 	stmt, err := parser.New().ParseOneStmt("select * from t where a > ?", "", "")
 	require.NoError(t, err)
-	_, fuzzyDigest := bindinfo.NormalizeStmtForFuzzyBinding(stmt)
+	_, fuzzyDigest := norm.NormalizeStmtForFuzzyBinding(stmt)
 	bindData, err := dom.BindHandle().MatchGlobalBinding(tk.Session(), fuzzyDigest, bindinfo.CollectTableNames(stmt))
 	require.NoError(t, err)
 	require.NotNil(t, bindData)
@@ -511,7 +512,7 @@ func TestHintsSetID(t *testing.T) {
 
 	internal.UtilCleanBindingEnv(tk, dom)
 	tk.MustExec("create global binding for select * from t where a > 10 using select /*+ use_index(t, idx_a) */ * from t where a > 10")
-	_, fuzzyDigest = bindinfo.NormalizeStmtForFuzzyBinding(stmt)
+	_, fuzzyDigest = norm.NormalizeStmtForFuzzyBinding(stmt)
 	bindData, err = dom.BindHandle().MatchGlobalBinding(tk.Session(), fuzzyDigest, bindinfo.CollectTableNames(stmt))
 	require.NoError(t, err)
 	require.NotNil(t, bindData)
@@ -522,7 +523,7 @@ func TestHintsSetID(t *testing.T) {
 
 	internal.UtilCleanBindingEnv(tk, dom)
 	tk.MustExec("create global binding for select * from t where a > 10 using select /*+ use_index(@sel_1 t, idx_a) */ * from t where a > 10")
-	_, fuzzyDigest = bindinfo.NormalizeStmtForFuzzyBinding(stmt)
+	_, fuzzyDigest = norm.NormalizeStmtForFuzzyBinding(stmt)
 	bindData, err = dom.BindHandle().MatchGlobalBinding(tk.Session(), fuzzyDigest, bindinfo.CollectTableNames(stmt))
 	require.NoError(t, err)
 	require.NotNil(t, bindData)
@@ -533,7 +534,7 @@ func TestHintsSetID(t *testing.T) {
 
 	internal.UtilCleanBindingEnv(tk, dom)
 	tk.MustExec("create global binding for select * from t where a > 10 using select /*+ use_index(@qb1 t, idx_a) qb_name(qb1) */ * from t where a > 10")
-	_, fuzzyDigest = bindinfo.NormalizeStmtForFuzzyBinding(stmt)
+	_, fuzzyDigest = norm.NormalizeStmtForFuzzyBinding(stmt)
 	bindData, err = dom.BindHandle().MatchGlobalBinding(tk.Session(), fuzzyDigest, bindinfo.CollectTableNames(stmt))
 	require.NoError(t, err)
 	require.NotNil(t, bindData)
@@ -544,7 +545,7 @@ func TestHintsSetID(t *testing.T) {
 
 	internal.UtilCleanBindingEnv(tk, dom)
 	tk.MustExec("create global binding for select * from t where a > 10 using select /*+ use_index(T, IDX_A) */ * from t where a > 10")
-	_, fuzzyDigest = bindinfo.NormalizeStmtForFuzzyBinding(stmt)
+	_, fuzzyDigest = norm.NormalizeStmtForFuzzyBinding(stmt)
 	bindData, err = dom.BindHandle().MatchGlobalBinding(tk.Session(), fuzzyDigest, bindinfo.CollectTableNames(stmt))
 	require.NoError(t, err)
 	require.NotNil(t, bindData)
@@ -557,7 +558,7 @@ func TestHintsSetID(t *testing.T) {
 	err = tk.ExecToErr("create global binding for select * from t using select /*+ non_exist_hint() */ * from t")
 	require.True(t, terror.ErrorEqual(err, parser.ErrParse))
 	tk.MustExec("create global binding for select * from t where a > 10 using select * from t where a > 10")
-	_, fuzzyDigest = bindinfo.NormalizeStmtForFuzzyBinding(stmt)
+	_, fuzzyDigest = norm.NormalizeStmtForFuzzyBinding(stmt)
 	bindData, err = dom.BindHandle().MatchGlobalBinding(tk.Session(), fuzzyDigest, bindinfo.CollectTableNames(stmt))
 	require.NoError(t, err)
 	require.NotNil(t, bindData)
