+++ pkg/planner/core/BUILD.bazel
@@ -87,6 +87,7 @@ go_library(
     deps = [
         "//br/pkg/storage",
         "//pkg/bindinfo",
+        "//pkg/bindinfo/norm",
         "//pkg/config",
         "//pkg/distsql",
         "//pkg/domain",
